Keep runner module under 400 lines by extracting reporter builder
Moving the ninja convenience wrappers into the runner boundary pushed
src/runner/mod.rs to 436 lines, over the Whitaker module_max_lines limit
of 400. Extract reporter construction (ReporterOptions, make_reporter,
should_force_text_task_updates) into a dedicated reporter.rs submodule and
move its unit test alongside, returning the module to 394 lines without
behaviour changes.
