+++ src/runner/tests.rs
@@ -21,19 +21,3 @@ fn resolve_output_path_respects_directory(
     let resolved = resolve_output_path(&cli, Path::new(input));
     assert_eq!(resolved.as_ref(), Path::new(expected));
 }
-
-#[rstest]
-#[case(OutputMode::Standard, true, false)]
-#[case(OutputMode::Standard, false, true)]
-#[case(OutputMode::Accessible, true, true)]
-#[case(OutputMode::Accessible, false, true)]
-fn force_text_task_updates_when_required(
-    #[case] mode: OutputMode,
-    #[case] stdout_is_tty: bool,
-    #[case] expected: bool,
-) {
-    assert_eq!(
-        should_force_text_task_updates(mode, stdout_is_tty),
-        expected
-    );
-}
